// packages/universe/src/galaxy/skyCalibration.generated.ts
// Generated by apps/ingest/src/galaxyReference.ts from GAMBONS RadianceOut.csv.
export const SKY_REGION_REFERENCES = [
  {
    name: 'Mid-latitudes',
    longitude: [0, 360],
    absoluteLatitude: [30, 60],
    pixels: 288616,
    vNanowatts: 53.5588014,
    photopicNanowatts: 65.4574786,
  },
  {
    name: 'Polar caps',
    longitude: [0, 360],
    absoluteLatitude: [60, 90],
    pixels: 105624,
    vNanowatts: 41.8620565,
    photopicNanowatts: 50.0140282,
  },
  {
    name: 'Aquila sightline',
    longitude: [40, 50],
    absoluteLatitude: [0, 5],
    pixels: 1910,
    vNanowatts: 176.410218,
    photopicNanowatts: 246.899429,
  },
] as const